Updating Nuget

+++ src/Comet/Helpers/ControlsExtensions.cs
@@ -4,7 +4,7 @@ namespace Comet
 {
 	public static class ControlsExtensions
 	{
-		public static float GetPercent(this Slider slider)
+		public static double GetPercent(this Slider slider)
 		{
 			var end = slider.Through?.CurrentValue ?? 0;
 			var current = slider.Value?.CurrentValue ?? 0;
